Updates to Arduino Code
Set up a definitions file for the MRSKULL pattern arrays.

Added a second mode, although I forget the name.. Biting Skull?
Whatever, it is two frames, one with the mouth open and the other with
mouth closed.

Played around with the brightness by removing the delay for blank pixels
(so that ON pixels were ON more ONften. Then set the leds to stay lit
longer. This all worked so now it's time to test the battery life at
this increased brightness level.

Again, I'm pretty exhausted after spending another night grinding away
on matrixes. It is definately time for a break and some extended periods
of sleep time goodness. Wild Stallions Rule!!